refactor : entity 패키지 생성 (UserPoint 및 PointHistory 이동)
