+++ src/main/java/io/hhplus/tdd/database/PointHistoryTable.java
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.database;
 
 
-import io.hhplus.tdd.point.PointHistory;
+import io.hhplus.tdd.point.entity.PointHistory;
 import io.hhplus.tdd.point.domain.TransactionType;
 import org.springframework.stereotype.Component;
