+++ src/main/java/io/hhplus/tdd/point/entity/UserPoint.java
@@ -1,7 +1,6 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.entity;
 
 import io.hhplus.tdd.point.domain.UserPointDomain;
-import lombok.Builder;
 
 public record UserPoint(
         long id,
