+++ src/main/java/io/hhplus/tdd/point/repository/PointHistoryRepository.java
@@ -2,8 +2,8 @@ package io.hhplus.tdd.point.repository;
 
 
 import io.hhplus.tdd.database.PointHistoryTable;
-import io.hhplus.tdd.point.PointHistory;
 import io.hhplus.tdd.point.domain.TransactionType;
+import io.hhplus.tdd.point.entity.PointHistory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
