+++ src/main/java/io/hhplus/tdd/point/service/PointService.java
@@ -1,7 +1,6 @@
 package io.hhplus.tdd.point.service;
 
-import io.hhplus.tdd.point.PointHistory;
-import io.hhplus.tdd.point.UserPoint;
+import io.hhplus.tdd.point.entity.PointHistory;
 import io.hhplus.tdd.point.controller.dto.RequestDto;
 import io.hhplus.tdd.point.domain.PointHistoryDomain;
 import io.hhplus.tdd.point.domain.UserPointDomain;
